+++ src/test/game/cands/test_development_hook.cpp
@@ -69,8 +69,6 @@ bool evaluate_test_case_development_hook(const StringTestCase *tc) {
         strings.push_back(sqtos(gs.frames[develop_hook.id][i].centre));
     }
 
-    // print_feature_frames(gs.feature_frames[0]);
-
     return assert_string_lists_equal(strings, tc->expected_results);
 }
