+++ src/test/game/cands/test_knight_fork_hook.cpp
@@ -121,8 +121,6 @@ bool evaluate_test_case_knight_fork_hook(const StringTestCase *tc) {
         }
     }
 
-    // print_feature_frames(gs.feature_frames[0]);
-
     return assert_string_lists_equal(strings, tc->expected_results);
 }
