+++ src/test/game/cands/test_unsafe_piece_hook.cpp
@@ -107,8 +107,6 @@ bool evaluate_test_case_unsafe_piece_hook(const StringTestCase *tc) {
         strings.push_back(sqtos(ff->centre));
     }
 
-    // print_feature_frames(gs.feature_frames[0]);
-
     return assert_string_lists_equal(strings, tc->expected_results);
 }
